I can only go one way, I have got no reverse gear

+++ 0x05-pointers_arrays_strings/2-strlen.c
@@ -5,7 +5,7 @@
  * @str: the string to get the length of
  * Return: The length of @str
  */
-size_t _strlen(const char *str)
+int _strlen(const char *str)
 {
 	size_t length = 0;
